Fixed sitemap and Prettier config
