+++ .dockerignore
@@ -1,0 +1,19 @@
+# Root folder of this turbo monorepo
+# Doesn't send useless context to Docker daemon
+.git
+.github
+about
+apps
+docker
+examples
+i18n
+**/node_modules
+spec
+supabase
+tests
+
+**/.next
+.vercel
+.env.*
+
+.dockerignore
