+++ .prettierrc
@@ -1,0 +1,11 @@
+{
+  "trailingComma": "es5",
+  "tabWidth": 2,
+  "semi": false,
+  "singleQuote": true,
+  "printWidth": 100,
+  "endOfLine": "lf",
+  "sqlKeywordCase": "lower",
+  "pluginSearchDirs": false,
+  "plugins": []
+}
