--- assets/js/generate-sitemap.mjs
@@ -1,56 +1,0 @@
-import { writeFileSync } from 'fs';
-import { globby } from 'globby';
-// import prettier from 'prettier';
-
-async function generate() {
-//   const prettierConfig = await prettier.resolveConfig('./.prettierrc.js');
-
-  const pages = await globby([
-    // 'pages/*.js',
-    // 'data/**/*.mdx',
-    // '!data/*.mdx',
-    // '!pages/_*.js',
-    // '!pages/api',
-    // '!pages/404.js',
-	'app/page.tsx',
-	'app/**/page.tsx'
-  ]);
-
-  const sitemap =
-`<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-        ${pages
-          .map((page) => {
-            const path = page
-            //   .replace('pages', '')
-            //   .replace('data', '')
-            //   .replace('.js', '')
-            //   .replace('.mdx', '');
-				.replace('app', '')
-				.replace('.tsx', '')
-				.replace('page', '');
-            const route = path === '/page' ? '' : path;
-
-            return `
-	<url>
-		<loc>${`https://www.stoneydsp.com${route}`}</loc>
-	</url>`;
-          })
-          .join('')}
-</urlset>
-    `;
-
-//   const formatted = prettier.format(sitemap, {
-//     ...prettierConfig,
-//     parser: 'html',
-//   });
-
-//   // eslint-disable-next-line no-sync
-//   writeFileSync('public/sitemap.xml', formatted);
-
-  // eslint-disable-next-line no-sync
-  writeFileSync('public/sitemap.xml', sitemap);
-
-}
-
-generate();
