+++ public/www_sitemap.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url>
+    <loc>https://stoneydsp.com/</loc>
+  </url>
+  <url>
+    <loc>https://stoneydsp.com/account/</loc>
+  </url>
+  <url>
+    <loc>https://stoneydsp.com/countries/</loc>
+  </url>
+  <url>
+    <loc>https://stoneydsp.com/example/</loc>
+  </url>
+  <url>
+    <loc>https://stoneydsp.com/login/</loc>
+  </url>
+</urlset>
